Loading spinner for submission history table
